+++ src/Bundles/Spelling/Exception/SpellingFixingNotSupportedException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace PHPUnuhi\Bundles\Spelling\Exception;
+
+class SpellingFixingNotSupportedException extends \Exception
+{
+
+}