+++ src/Bundles/Spelling/Exception/SpellingValidationNotSupportedException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace PHPUnuhi\Bundles\Spelling\Exception;
+
+class SpellingValidationNotSupportedException extends \Exception
+{
+
+}